+++ src/modules/Sales/sales.repository.ts
@@ -1,0 +1,6 @@
+import { Repository, EntityRepository } from 'typeorm';
+
+import { SalesEntity } from './sales.entity';
+
+@EntityRepository(SalesEntity)
+export class SalesRepository extends Repository<SalesEntity> {}
