--- store-back/Dockerfile
@@ -1,6 +1,0 @@
-FROM node:12
-WORKDIR /app
-COPY package*.json ./
-RUN npm i
-COPY . .
-CMD ["npm" ,"start"]
