--- store-back/src/config/config.keys.ts
@@ -1,8 +1,0 @@
-export enum Configuration {
-  PORT = 'PORT',
-  HOST = 'HOST',
-  USERNAME = 'USERNAME',
-  PASSWORD = 'PASSWORD',
-  DATABASE = 'DATABASE',
-  JWT_SECRET = 'JWT_SECRET',
-}
