--- store-back/src/config/config.module.ts
@@ -1,13 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ConfigService } from './config.service';
-
-@Module({
-  providers: [
-    {
-      provide: ConfigService,
-      useValue: new ConfigService(),
-    },
-  ],
-  exports: [ConfigService],
-})
-export class ConfigModule {}
