--- store-back/src/database/database.service.ts
@@ -1,29 +1,0 @@
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { ConfigModule } from 'src/config/config.module';
-import { ConfigService } from 'src/config/config.service';
-import { ConnectionOptions } from 'typeorm';
-import { Configuration } from 'src/config/config.keys';
-
-export const databaseProviders = [
-  TypeOrmModule.forRootAsync({
-    imports: [ConfigModule],
-    inject: [ConfigService],
-    async useFactory(config: ConfigService) {
-      return {
-        ssl: false,
-        type: 'postgres' as 'postgres',
-        database: config.get(Configuration.DATABASE),
-        host: config.get(Configuration.HOST),
-        username: config.get(Configuration.USERNAME),
-        password: config.get(Configuration.PASSWORD),
-        entities: [__dirname + '/../**/*.entity{.ts,.js}'],
-        migrations: [__dirname + '/migrations/*{.ts,.js}'],
-        migrationsRun: true,
-        synchronize: true,
-        cli: {
-          migrationsDir: 'src/database/migrations',
-        },
-      } as ConnectionOptions;
-    },
-  }),
-];
