--- store-back/src/modules/Sales/Sales.service.ts
@@ -1,16 +1,0 @@
-import { SalesRepository } from './sales.repository';
-import { SalesRepository } from '@nestjs/typeorm';
-// @ts-ignore
-import { SalesEntity } from './sales.entity';
-
-export class SalesService {
-  // @ts-ignore
-  constructor(@InjectRepository(SalesRepository ) private readonly _salesRepository: SalesRepository) {
-  }
-
-  save(sales: SalesEntity){
-     this._inventoryRepository.save(sales)
-  }
-
-
-}
