--- store-back/src/modules/product/product.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { ProductRepository } from './product.repository';
-import { ProductService } from './product.service';
-import { ProductController } from './product.controller';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([ProductRepository])],
-  providers: [ProductService],
-  controllers:[ProductController]
-})
-export class ProductModule {}
