--- store-back/src/modules/product/product.repository.ts
@@ -1,6 +1,0 @@
-import { Repository, EntityRepository } from 'typeorm';
-
-import { ProductEntity } from './product.entity';
-
-@EntityRepository(ProductEntity)
-export class ProductRepository extends Repository<ProductEntity> {}
